// LDC.Atlas.WebApp/Client/app/shared/entities/parent-costs-to-adjust.entity.ts
export class ParentCostsToAdjust {
    costId: number;
    sectionId: number;
    costTypeCode: string;
    supplierCode: string;
    costDirection: string;
    currencyCode: string;
    rateTypeId: number;
    rateTypeCode: string;
    priceUnitId: number;
    priceCode: string;
    costDirectionId: number;
    rate: number;
    narrative: string;
    isProRated: boolean;
    isOverRide: boolean;
    quantity: number;
    invoicingStatusId: number;
}